% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_vc.R
\name{calculate_vcs}
\alias{calculate_vcs}
\title{Extracts varying coefficient estimates (for SVC, TVC and STVC)}
\usage{
calculate_vcs(model, terms, data)
}
\arguments{
\item{model}{a GAM model with smooths created using the \code{mgcv} package}

\item{terms}{a vector of names starting with "Intercept" plus the names of the covariates used in the GAM model (these are the names of the variables in \code{data} )}

\item{data}{the data used to create the GAM model, \code{data.frame} or \code{tibble} format}
}
\value{
a \code{data.frame} of the input data and the coefficient and standard error estimates for each covariate
}
\description{
Extracts varying coefficient estimates (for SVC, TVC and STVC)
}
