% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_bma.R
\name{do_bma}
\alias{do_bma}
\title{Undertake undertake coefficient averaging using Bayesian Model Avergaing (BMA), weighting different models by their probabailities}
\usage{
do_bma(model_table, terms, thresh = 0.1, relative = F, data)
}
\arguments{
\item{model_table}{a table of competing models generated by \code{gam_model_probs}}

\item{terms}{a vector of names starting with "Intercept" plus the names of the covariates used in the GAM model (these are the names of the variables in  \code{data} )}

\item{thresh}{a probability threshold value above which to combine competing models}

\item{relative}{a logical to indicate whether the probabilities in \code{data} are relative (\code{Pr(M)}) or absolute (\code{Pr(M|D)})}

\item{data}{the inupt data with a nmaed Intercept term}
}
\value{
a matrix of the probability weighted averaged coefficient estimates
}
\description{
Undertake undertake coefficient averaging using Bayesian Model Avergaing (BMA), weighting different models by their probabailities
}
