% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_model_scores.R
\name{gam_model_scores}
\alias{gam_model_scores}
\title{Title Ranks models by GCV, giving the model form for each predctor variable.}
\usage{
gam_model_scores(res_tab, n = 10)
}
\arguments{
\item{res_tab}{a \code{data.frame} returned from the \code{evaluate_models()} function.}

\item{n}{the number of ranked models to return.}
}
\value{
a \code{tibble} of the 'n' best models, ranked by GCV, with the form of each predictor variable where '---' indicates the absence of a predictor, 'Fixed' that a parametric form was specified,  'te_S' a spatial Tensor Product (TP) smooth, 'te_T'  a temporal TP smooth and 'te_ST' a spatio-temporal TP smooth.
}
\description{
Title Ranks models by GCV, giving the model form for each predctor variable.
}
\examples{
require(dplyr)
# define input data
input_data = productivity |> filter(year == 1975) |> mutate(Intercept = 1)
# detemine length ranges
rho_sp <- opt_length_scale(input_data,
       target_var = "privC",
       vars = c("Intercept", "unemp", "pubC"),
       coords_x = "X",
       coords_y = "Y",
       STVC = FALSE)
# evaluate different model forms
svc_mods = evaluate_models(
       input_data = input_data,
       target_var = "privC",
       vars = c("unemp", "pubC"),
       coords_x = "X",
       coords_y = "Y",
       STVC = FALSE,
       rho_space_vec = round(rho_sp$rho_space,1))
# rank models and translate predicor variable indices
mod_comp <- gam_model_scores(svc_mods)
# have a look
mod_comp
}
