% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lb}
\alias{lb}
\title{London borough boundaries}
\format{
A \code{sf} polygon (MULTIPOLYGON) dataset with 33 observations and 2 fields.
\describe{
\item{name}{The name of the London borough}
\item{lad}{The ONS lcoal authrority district code for the borough}
}
}
\source{
Lu, Binbin, Harris, Paul, Charlton, Martin, Brunsdon, Chris, Nakaya, Tomoki, Murakami, Daisuke, Hu, Yigong, Evans, Fiona H, Høglund, Hjalmar. 2024. Geographically-Weighted Models
}
\usage{
lb
}
\description{
A spatial dataset of of the boundaries of the 33 London Boroughs extracted from the \code{GWModel} package, cleaned and converted to \code{sf}.
}
\examples{
data("lb")
}
\keyword{datasets}
