% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_length_scale.R
\name{opt_length_scale}
\alias{opt_length_scale}
\title{Title Optimises the length scale parameters for Tensor Product smooths}
\usage{
opt_length_scale(
  input_data,
  target_var = "privC",
  vars = c("Intercept", "unemp", "pubC"),
  coords_x = "X",
  coords_y = "Y",
  STVC = FALSE,
  time_var = NULL,
  ...
)
}
\arguments{
\item{input_data}{the data to be used used to create the GAM model in (\code{data.frame} or \code{tibble} format), containing an Intercept column to allow it be treated as an addressable term in the model.}

\item{target_var}{the name of the target variable.}

\item{vars}{a vector of "Intercept" and the predictor variable names.}

\item{coords_x}{the name of the X, Easting or Longitude variable in \code{input_data}.}

\item{coords_y}{the name of the Y, Northing or Latitude variable in \code{input_data}.}

\item{STVC}{a logical operator indicating whether the model is space-time (\code{TRUE}) or just space (\code{FALSE}).}

\item{time_var}{the name of the time variable if undertaking STVC model evaluations}

\item{...}{other prameters to be passed to \code{mgcv} \code{gam()} function.}
}
\value{
A \code{data.frame} containing the spatial and temporal length scales \eqn{\rho_{space}} and \eqn{\rho_{time}}) for use in the Tensor Product smooths.
}
\description{
Title Optimises the length scale parameters for Tensor Product smooths
}
\examples{
require(dplyr)
# define input data
input_data = productivity |> filter(year == 1975) |> mutate(Intercept = 1)
# determine length ranges
rho_sp <- opt_length_scale(input_data,
       target_var = "privC",
       vars = c("Intercept", "unemp", "pubC"),
       coords_x = "X",
       coords_y = "Y",
       STVC = FALSE)
# have a look!
rho_sp
}
