% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_1d_smooth.R
\name{plot_1d_smooth}
\alias{plot_1d_smooth}
\title{Plots a 1-Dimensional GAM smooth}
\usage{
plot_1d_smooth(mod, ncol = NULL, nrow = NULL, fills = "lightblue")
}
\arguments{
\item{mod}{a GAM model with smooths created using the mgcv package}

\item{ncol}{the number of columns for the compound plot}

\item{nrow}{the number of rows for the compound plot}

\item{fills}{the fill colours (single or vector)}
}
\value{
a compound plot of the 1D smooths (rendered using cowplot::plot_grid )
}
\description{
Plots a 1-Dimensional GAM smooth
}
\examples{
library(mgcv)
library(tidyverse)
library(cowplot)
set.seed(2) ## simulate some data...
dat <- gamSim(1,n=400,dist="normal",scale=2)
b <- gam(y~s(x0)+s(x1)+s(x2)+s(x3),data=dat)
plot_1d_smooth(b, ncol = 2, fills = c("lightblue", "lightblue3"))
}
