% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{productivity}
\alias{productivity}
\title{US States Economic Productivity Data (1970-1985)}
\format{
A tibble with 816 rows and 14 columns.
\describe{
\item{state}{The name of the state}
\item{GEOID}{The state code}
\item{region}{The region}
\item{pubC}{Public capital which is composed of highways and streets (hwy) water and sewer facilities (water) and other public buildings and structures (util)}
\item{hwy}{Highway and streets assets}
\item{water}{Water utility assets}
\item{util}{Other public buildings and structures}
\item{privC}{Private captial stock}
\item{gsp}{Gross state product}
\item{emp}{Labour input measured by the employment in non-agricultural payrolls}
\item{unemp}{State unemployment rate capture elements of the business cycle}
\item{X}{Easting in metres from USA Contiguous Equidistant Conic projection (ESRI:102005)}
\item{Y}{Northing in metres from USA Contiguous Equidistant Conic projection (ESRI:102005)}
}
}
\source{
Croissant, Yves, Giovanni Millo, and Kevin Tappe. 2022. Plm: Linear Models for Panel Data
}
\usage{
productivity
}
\description{
A dataset of annual economic productivity data for the 48 contiguous US states (with Washington DC merged into Maryland), from 1970 to 1985 (17 years) in long format. The data productivity data table was extracted from the \code{plm} package.
}
\examples{
data(productivity)
}
\keyword{datasets}
