% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_vcs.R
\name{calculate_vcs}
\alias{calculate_vcs}
\title{Extracts varying coefficient estimates (for SVC, TVC and STVC models).}
\usage{
calculate_vcs(input_data, model, terms)
}
\arguments{
\item{input_data}{the data used to create the GAM model in \code{data.frame}, \code{tibble} or \code{sf} format}

\item{model}{a GAM model with smooths created using the \code{mgcv} package}

\item{terms}{a vector of names starting with "Intercept" plus the names of the covariates used in the GAM model (these are the names of the variables in \code{data} )}
}
\value{
A \code{data.frame} of the input data and the coefficient and standard error estimates for each covariate.
}
\description{
Extracts varying coefficient estimates (for SVC, TVC and STVC models).
}
\examples{
require(dplyr)
# define input data
input_data = productivity |> filter(year == 1975) |> mutate(Intercept = 1)
# detemine length ranges
rho_sp <- opt_length_scale(input_data,
       target_var = "privC",
       vars = c("Intercept", "unemp", "pubC"),
       coords_x = "X",
       coords_y = "Y",
       STVC = FALSE)
# evaluate different model forms
svc_mods = evaluate_models(
       input_data = input_data,
       target_var = "privC",
       vars = c("unemp", "pubC"),
       coords_x = "X",
       coords_y = "Y",
       STVC = FALSE,
       rho_space_vec = round(rho_sp$rho_space,1))
# rank models and translate predicor variable indices
mod_comp <- gam_model_scores(svc_mods)
# have a look
mod_comp
# select best model
f = as.formula(mod_comp$f[1])
# put into a `mgcv` GAM model
m = gam(f, data = input_data)
# calculate the Varying Coefficients
terms = c("Intercept", "unemp", "pubC")
vcs = calculate_vcs(input_data, m, terms)
vcs |> select(state, year, starts_with(c("b_", "se_")))
}
