% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_size.R
\name{effect_size}
\alias{effect_size}
\title{Quantifies the relative effect sizes of each component of zmgcv` GAM model.}
\usage{
effect_size(mgcv_model, digits = 3)
}
\arguments{
\item{mgcv_model}{a GAM model created by the \code{mgcv} package}

\item{digits}{the number of significant figures of uysed to report effect size}
}
\value{
a \code{tibble} of the 'n' best models, ranked by GCV, with the form of each predictor variable where '---' indicates the absence of a predictor, 'Fixed' that a parametric form was specified,  's_S' a spatial smooth, 's_T'  a temporal smooth and 'te_ST' a spatio-temporal smooth.
}
\description{
Quantifies the relative effect sizes of each component of zmgcv` GAM model.
}
\examples{
require(dplyr)
require(stringr)
require(purrr)
require(doParallel)

# define input data
data("chaco")
m <- gam(
ndvi ~
  te(X,Y,month, by = tmax, bs = c("tp", "cr"), d = c(2,1)) +
  pr + s(X,Y, by = pr) + s(month, by = pr),
 data = chaco,
 method = "REML",
 family = gaussian()
)
# examine the effect size
effect_size(m, 3)
}
