% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_models.R
\name{evaluate_models}
\alias{evaluate_models}
\title{Creates evaluates multiple varying coefficient GAM GP smooth models (SVC or STVC)}
\usage{
evaluate_models(
  data,
  target_var = "privC",
  covariates = c("unemp", "pubC"),
  coords_x = "X",
  coords_y = "Y",
  STVC = TRUE,
  time_var = NULL
)
}
\arguments{
\item{data}{a \code{data.frame} or \code{tibble} containing the target variables, covariates and coordinate variables}

\item{target_var}{the name of the target variable in \code{data}}

\item{covariates}{the name of the covariates (predictor variables) in \code{data}}

\item{coords_x}{the name of the X, Easting or Longitude variable in \code{data}}

\item{coords_y}{the name of the Y, Northing or Latitude variable in \code{data}}

\item{STVC}{a logical operator to indicate whether the models Space-Time (\code{TRUE}) or just Space (\code{FALSE})}

\item{time_var}{the name of the time variable if undertaking STVC model evaluations}
}
\value{
a data table in data.frame format of all possible model combinations with each covraite specified in all possible ways
}
\description{
Creates evaluates multiple varying coefficient GAM GP smooth models (SVC or STVC)
}
\examples{
library(glue)
library(purrr)
library(doParallel)
library(mgcv)
data(productivity)
data = productivity |> filter(year == "1970")
svc_res_gam = evaluate_models(data, STVC = FALSE)
head(svc_res_gam)
}
