% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_model_probs.R
\name{gam_model_probs}
\alias{gam_model_probs}
\title{Calculates the \verb{model probabilities of the different GAM models generated by }evaluate_models`}
\usage{
gam_model_probs(res_tab, n = 10)
}
\arguments{
\item{res_tab}{a table generated by \code{evaluate_models}}

\item{n}{the number of models to retain and generate probabilities for}
}
\value{
a ranked data table in \code{tibble} format of the top \code{n} models, their form, BIC and probability or relative probability
}
\description{
Calculates the \verb{model probabilities of the different GAM models generated by }evaluate_models`
}
