% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stgam-package.R
\docType{package}
\name{stgam-package}
\alias{stgam}
\alias{stgam-package}
\title{stgam: Spatially and Temporally Varying Coefficient Models Using Generalized Additive Models}
\description{
A framework for specifying spatially, temporally and spatially-and-temporally varying coefficient models using Generalized Additive Models (GAMs) and Tensor Product (TP) smooths with Gaussian Process bases. It builds GAM functionality form the `mgcv` package. The TP smooths are parameterised with location and time attributes, and the length scales (range parameters) are optimised for each predictor variable, providing an explicit indication of the scales of interaction with the target variable. In this way the framework supports the investigation of the presence and nature of any space-time dependencies in the data by evaluating multiple model forms (specifications) using a Generalized Cross-Validation (GCV) score. The workflow sequence is i) determine length ranges with `opt_length_scale`, ii) evaluate different model forms with `evaluate_models`, iii) to rank models and translate predictor variable indices with `gam_model_scores` and pick the best model, abd iv) to calculate the varying coefficient estimates. For more details see: Comber et al (2023) \doi{10.4230/LIPIcs.GIScience.2023.22}, Comber et al (2024) \doi{10.1080/13658816.2023.2270285} and Comber et al (2004) \doi{10.3390/ijgi13120459}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lexcomber/stgam}
  \item Report bugs at \url{https://github.com/lexcomber/stgam/issues}
}

}
\author{
\strong{Maintainer}: Lex Comber \email{a.comber@leeds.ac.uk}

Other contributors:
\itemize{
  \item Paul Harris \email{paul.harris@rothamsted.ac.uk} [contributor]
  \item Chris Brunsdon \email{christopher.brunsdon@mu.ie} [contributor]
}

}
\keyword{internal}
